requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==1.10.7
llama-index==0.8.29.post1
openai==0.28.0
langchain==0.0.295